Split apart MemoryFaction
